rustdoc: Allow private modules be included in docs
Added a --all-modules/-a flag which causes rustdoc to include
documentation information even for modules which are not visible outside
of the crate they are defined in.

This flag is off by default. It is most useful if the strip-private pass
is disable but can be useful even without this. There is no change in
rustdoc's output with this flag off.